requirements.txt
pydantic-ai[mcp]
python-dotenv
langgraph